% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred-data.R
\docType{data}
\name{fred.qd}
\alias{fred.qd}
\title{FRED-QD Database for Macroeconomic Research}
\format{
A quarterly multivariate time series of the class mts.
}
\source{
Federal Reserve Bank of St. Louis, Michael McCracken and Serena Ng,
\href{https://www.stlouisfed.org/research/economists/mccracken/fred-databases}{Source},
License \href{https://cran.r-project.org/web/packages/BVAR/LICENSE}{ODC-BY 1.0}
}
\usage{
data(fred.qd)
}
\description{
FRED-MD and FRED-QD are large macroeconomic databases, containing monthly
and quarterly time series of US macro variables that are frequently used in the literature. They
are intended to facilitate the reproduction of empirical work and simplify
data related tasks.
For variable descriptions and transformation codes see McCracken and Ng (2016) or
\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}.
}
\details{
The versions of FRED-MD and FRED-QD that are provided here are taken from
the `BVAR` package available
\href{https://cran.r-project.org/web/packages/BVAR/index.html}{here}.
The data is licensed  under a modified ODC-BY 1.0 license that can be found
in the
\href{https://cran.r-project.org/web/packages/BVAR/LICENSE}{LICENSE file}
of the `BVAR` package.
We would like to thank Michael McCracken and Serena Ng, Adrienne Brennecke
and the Federal Reserve Bank of St. Louis for creating, updating and making
available the datasets and many of the contained time series. We also thank
all other owners of included time series that permitted their use.
Furthermore, we thank the authors of the
\href{https://cran.r-project.org/web/packages/BVAR/index.html}{BVAR} package
for making the database available under the ODC-BY 1.0 license.
}
\references{
McCracken, M. W. and Ng, S. (2016) FRED-MD: A Monthly Database for
  Macroeconomic Research. \emph{Journal of Business & Economic Statistics},
  \bold{34:4}, 574-589, \doi{10.1080/07350015.2015.1086655}.
  McCracken, M. W., & Ng, S. (2020). FRED-QD: A Quarterly Database for
  Macroeconomic Research \bold{w26872}. National Bureau of Economic Research.
}
\keyword{datasets}
