% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destatis-data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{German Quarterly GDP}
\format{
A quarterly time series of the class ts.
}
\source{
Statistisches Bundesamt (Destatis), Genesis-Online, Tablename 81000BV007, License \href{https://www.govdata.de/dl-de/by-2-0}{DL-DE BY-2.0}
}
\usage{
data(gdp)
}
\description{
German Quarterly GDP
}
\keyword{datasets}
