% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd-data.R
\docType{data}
\name{temperature.y}
\alias{temperature.y}
\title{Yearly Average Temperature Germany}
\format{
A yearly time series of the class ts.
}
\source{
Deutscher Wetterdienst (DWD), \href{https://www.dwd.de/DE/leistungen/zeitreihen/zeitreihen.htm}{Source}, \href{https://www.dwd.de/DE/service/copyright/copyright_node.html}{Legal notice}, License \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0}
}
\usage{
data(temperature.y)
}
\description{
Yearly Average Temperature Germany
}
\keyword{datasets}
